Fix MPI_ANY_SOURCE error case when returning MPI_ERR_PENDING.
This required a change to ompi_errhandler_request_invoke() to work around
pending messages like this and return the correct error code. See how it is
used in ompi/mpi/c/wait.c and you will notice that the completion operations in
Open MPI ignore the return code from the internal operation, and rely
completely on the return code from the ompi_errhandler_request_invoke()
operation.